machine_learning/requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
joblib
openpyxl
django 
djangorestframework